feat: introduce package for runfile helpers
Introduces a new package called `@bazel/runfiles` that
provides runfile helpers for Bazel with NodeJS.

The benefit of providing the helpers as a separate package
is that people can explicitly specify a dependency on
the helpers and also can get typings. The current best practice
of using a require w/ an environment variable is untyped and
prone to mistakes (and doesn't fit well into the Node/NPM ecosystem).
