+++ BUILD.bazel
@@ -89,6 +89,7 @@ pkg_npm(
         "//internal/pkg_npm:package_contents",
         "//internal/pkg_web:package_contents",
         "//internal/providers:package_contents",
+        "//internal/runfiles:package_contents",
         "//third_party/github.com/bazelbuild/bazel:package_contents",
         "//third_party/github.com/bazelbuild/bazel-skylib:package_contents",
         "//third_party/github.com/bazelbuild/bazel/tools/bash/runfiles:package_contents",
