+++ examples/create-react-app/BUILD.bazel
@@ -59,7 +59,10 @@ react_scripts(
 
 nodejs_test(
     name = "build_smoke_test",
-    data = ["build"],
+    data = [
+        "build",
+        "@npm//@bazel/runfiles",
+    ],
     entry_point = "build_smoke_test.js",
 )
