+++ internal/linker/link_node_modules.ts
@@ -6,12 +6,16 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
+// We cannot rely from the linker on the `@bazel/runfiles` package, hence we import from
+// the runfile helper through a checked-in file from `internal/runfiles`. In order to still
+// have typings we use a type-only import to the `@bazel/runfiles` package that is the source
+// of truth for the checked-in file.
+const {runfiles: _defaultRunfiles, _BAZEL_OUT_REGEX}:
+    typeof import('@bazel/runfiles') = require('../runfiles/index.js')
+import {Runfiles} from '@bazel/runfiles';
+
 // Run Bazel with --define=VERBOSE_LOGS=1 to enable this logging
 const VERBOSE_LOGS = !!process.env['VERBOSE_LOGS'];
-
-// NB: on windows thanks to legacy 8-character path segments it might be like
-// c:/b/ojvxx6nx/execroot/build_~1/bazel-~1/x64_wi~1/bin/internal/npm_in~1/test
-const BAZEL_OUT_REGEX = /(\/bazel-out\/|\/bazel-~1\/x64_wi~1\/)/;
 
 function log_verbose(...m: string[]) {
   if (VERBOSE_LOGS) console.error('[link_node_modules.js]', ...m);
@@ -19,17 +23,6 @@ function log_verbose(...m: string[]) {
 
 function log_error(error: Error) {
   console.error('[link_node_modules.js] An error has been reported:', error, error.stack);
-}
-
-function panic(m: string) {
-  throw new Error(`Internal error! Please run again with
-   --define=VERBOSE_LOG=1
-and file an issue: https://github.com/bazelbuild/rules_nodejs/issues/new?template=bug_report.md
-Include as much of the build output as you can without disclosing anything confidential.
-
-  Error:
-  ${m}
-  `);
 }
 
 /**
@@ -181,177 +174,6 @@ async function resolveExternalWorkspacePath(
     // so we must symlink into `runfiles/npm/node_modules`. This directory exists whether legacy
     // external runfiles are on or off.
     return path.resolve(`${startCwd}/../${workspace}`)
-  }
-}
-export class Runfiles {
-  manifest: Map<string, string>|undefined;
-  dir: string|undefined;
-  /**
-   * If the environment gives us enough hints, we can know the workspace name
-   */
-  workspace: string|undefined;
-  /**
-   * If the environment gives us enough hints, we can know the package path
-   */
-  package: string|undefined;
-
-  constructor(env: typeof process.env) {
-    // If Bazel sets a variable pointing to a runfiles manifest,
-    // we'll always use it.
-    // Note that this has a slight performance implication on Mac/Linux
-    // where we could use the runfiles tree already laid out on disk
-    // but this just costs one file read for the external npm/node_modules
-    // and one for each first-party module, not one per file.
-    if (!!env['RUNFILES_MANIFEST_FILE']) {
-      this.manifest = this.loadRunfilesManifest(env['RUNFILES_MANIFEST_FILE']!);
-    } else if (!!env['RUNFILES_DIR']) {
-      this.dir = path.resolve(env['RUNFILES_DIR']!);
-    } else {
-      panic(
-          'Every node program run under Bazel must have a $RUNFILES_DIR or $RUNFILES_MANIFEST_FILE environment variable');
-    }
-    // Under --noenable_runfiles (in particular on Windows)
-    // Bazel sets RUNFILES_MANIFEST_ONLY=1.
-    // When this happens, we need to read the manifest file to locate
-    // inputs
-    if (env['RUNFILES_MANIFEST_ONLY'] === '1' && !env['RUNFILES_MANIFEST_FILE']) {
-      log_verbose(`Workaround https://github.com/bazelbuild/bazel/issues/7994
-                 RUNFILES_MANIFEST_FILE should have been set but wasn't.
-                 falling back to using runfiles symlinks.
-                 If you want to test runfiles manifest behavior, add
-                 --spawn_strategy=standalone to the command line.`);
-    }
-    // Bazel starts actions with pwd=execroot/my_wksp or pwd=runfiles/my_wksp
-    this.workspace = env['BAZEL_WORKSPACE'] || undefined;
-    // If target is from an external workspace such as @npm//rollup/bin:rollup
-    // resolvePackageRelative is not supported since package is in an external
-    // workspace.
-    const target = env['BAZEL_TARGET'];
-    if (!!target && !target.startsWith('@')) {
-      // //path/to:target -> path/to
-      this.package = target.split(':')[0].replace(/^\/\//, '');
-    }
-  }
-
-  lookupDirectory(dir: string): string|undefined {
-    if (!this.manifest) return undefined;
-
-    let result: string|undefined;
-    for (const [k, v] of this.manifest) {
-      // Account for Bazel --legacy_external_runfiles
-      // which pollutes the workspace with 'my_wksp/external/...'
-      if (k.startsWith(`${dir}/external`)) continue;
-
-      // Entry looks like
-      // k: npm/node_modules/semver/LICENSE
-      // v: /path/to/external/npm/node_modules/semver/LICENSE
-      // calculate l = length(`/semver/LICENSE`)
-      if (k.startsWith(dir)) {
-        const l = k.length - dir.length;
-        const maybe = v.substring(0, v.length - l);
-        if (maybe.match(BAZEL_OUT_REGEX)) {
-          return maybe;
-        } else {
-          result = maybe;
-        }
-      }
-    }
-    return result;
-  }
-
-
-  /**
-   * The runfiles manifest maps from short_path
-   * https://docs.bazel.build/versions/master/skylark/lib/File.html#short_path
-   * to the actual location on disk where the file can be read.
-   *
-   * In a sandboxed execution, it does not exist. In that case, runfiles must be
-   * resolved from a symlink tree under the runfiles dir.
-   * See https://github.com/bazelbuild/bazel/issues/3726
-   */
-  loadRunfilesManifest(manifestPath: string) {
-    log_verbose(`using runfiles manifest ${manifestPath}`);
-
-    const runfilesEntries = new Map();
-    const input = fs.readFileSync(manifestPath, {encoding: 'utf-8'});
-
-    for (const line of input.split('\n')) {
-      if (!line) continue;
-      const [runfilesPath, realPath] = line.split(' ');
-      runfilesEntries.set(runfilesPath, realPath);
-    }
-
-    return runfilesEntries;
-  }
-
-  resolve(modulePath: string) {
-    if (path.isAbsolute(modulePath)) {
-      return modulePath;
-    }
-    const result = this._resolve(modulePath, undefined);
-    if (result) {
-      return result;
-    }
-    const e = new Error(`could not resolve module ${modulePath}`);
-    (e as any).code = 'MODULE_NOT_FOUND';
-    throw e;
-  }
-
-  _resolve(moduleBase: string, moduleTail: string|undefined): string|undefined {
-    if (this.manifest) {
-      const result = this.lookupDirectory(moduleBase);
-      if (result) {
-        if (moduleTail) {
-          const maybe = path.join(result, moduleTail || '');
-          if (fs.existsSync(maybe)) {
-            return maybe;
-          }
-        } else {
-          return result;
-        }
-      }
-    }
-    if (runfiles.dir) {
-      const maybe = path.join(runfiles.dir, moduleBase, moduleTail || '');
-      if (fs.existsSync(maybe)) {
-        return maybe;
-      }
-    }
-    const dirname = path.dirname(moduleBase);
-    if (dirname == '.') {
-      // no match
-      return undefined;
-    }
-    return this._resolve(dirname, path.join(path.basename(moduleBase), moduleTail || ''));
-  }
-
-  resolveWorkspaceRelative(modulePath: string) {
-    if (!this.workspace) {
-      throw new Error(
-          'workspace could not be determined from the environment; make sure BAZEL_WORKSPACE is set');
-    }
-    return this.resolve(path.posix.join(this.workspace, modulePath));
-  }
-
-  resolvePackageRelative(modulePath: string) {
-    if (!this.workspace) {
-      throw new Error(
-          'workspace could not be determined from the environment; make sure BAZEL_WORKSPACE is set');
-    }
-    // NB: this.package may be '' if at the root of the workspace
-    if (this.package === undefined) {
-      throw new Error(
-          'package could not be determined from the environment; make sure BAZEL_TARGET is set');
-    }
-    return this.resolve(path.posix.join(this.workspace, this.package, modulePath));
-  }
-
-  patchRequire() {
-    const requirePatch = process.env['BAZEL_NODE_PATCH_REQUIRE'];
-    if (!requirePatch) {
-      throw new Error('require patch location could not be determined from the environment');
-    }
-    require(requirePatch);
   }
 }
 
@@ -544,7 +366,7 @@ function findExecroot(startCwd: string): string|undefined {
   // determine the execroot `b/f/w` by finding the first instance of bazel-out.
   // NB: If we are inside nodejs_image or a nodejs_binary run manually there may be no execroot
   // found.
-  const bazelOutMatch = startCwd.match(BAZEL_OUT_REGEX);
+  const bazelOutMatch = startCwd.match(_BAZEL_OUT_REGEX);
   return bazelOutMatch ? startCwd.slice(0, bazelOutMatch.index) : undefined;
 }
 
@@ -763,7 +585,7 @@ export async function main(args: string[], runfiles: Runfiles) {
           if (runfiles.manifest && modulePath.startsWith(`${bin}/`)) {
             // Check for BAZEL_OUT_REGEX and not /${bin}/ since resolution
             // may be in the `/bazel-out/host` if cfg = "host"
-            if (!target.match(BAZEL_OUT_REGEX)) {
+            if (!target.match(_BAZEL_OUT_REGEX)) {
               const e = new Error(`could not resolve module ${runfilesPath} in output tree`);
               (e as any).code = 'MODULE_NOT_FOUND';
               throw e;
@@ -838,8 +660,6 @@ export async function main(args: string[], runfiles: Runfiles) {
   return code;
 }
 
-export const runfiles = new Runfiles(process.env);
-
 if (require.main === module) {
   if (Number(process.versions.node.split('.')[0]) < 10) {
     console.error(`ERROR: rules_nodejs linker requires Node v10 or greater, but is running on ${
@@ -850,7 +670,7 @@ if (require.main === module) {
   }
   (async () => {
     try {
-      process.exitCode = await main(process.argv.slice(2), runfiles);
+      process.exitCode = await main(process.argv.slice(2), _defaultRunfiles);
     } catch (e) {
       log_error(e);
       process.exitCode = 1;
