--- internal/linker/runfiles_helper.js
@@ -1,1 +1,0 @@
-module.exports = require('./index.js').runfiles;
