+++ internal/linker/test/link_node_modules.spec.ts
@@ -1,8 +1,15 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import * as linker from '../link_node_modules';
 import {LinkerAliases, LinkerTreeElement} from '../link_node_modules';
+
+// The linker is imported through it's checked-in file. We do this because the import to
+// the runfile helpers from the linker file will always resolve to the checked-in file too.
+const linker: typeof import('../link_node_modules') = require('../index.js');
+// As seen above, the linker file always loads the checked-in runfile helpers. We don't want
+// to have a mix of checked-in files and sources, so we import runfile helpers from the
+// checked-in file too, but use the types provided by the `@bazel/runfiles` package.
+const {Runfiles}: typeof import('@bazel/runfiles') = require('../../runfiles/index.js');
 
 const BIN_DIR = `bazel-out/my-platform-fastbuild/bin`;
 
@@ -378,7 +385,7 @@ describe('link_node_modules', () => {
     });
 
     // TODO(alexeagle): test should control the environment, not just pass through
-    await linker.main(['manifest.json'], new linker.Runfiles(process.env));
+    await linker.main(['manifest.json'], new Runfiles(process.env));
 
     // The linker expects to run as its own process, so it changes the wd
     process.chdir(path.join());
@@ -420,7 +427,7 @@ describe('link_node_modules', () => {
       roots: {}
     });
 
-    await linker.main(['manifest.json'], new linker.Runfiles({
+    await linker.main(['manifest.json'], new Runfiles({
       // This test assumes an environment where runfiles are not symlinked. Hence
       // we pass a runfile manifest file.
       'RUNFILES_MANIFEST_FILE': 'runfiles.mf',
@@ -449,7 +456,7 @@ describe('link_node_modules', () => {
     // In the second run, we added a mapping for `@angular/cdk`. This means
     // that the linker would need to clean up the previous `cdk/bidi` link
     // in order to be able to create a link for `@angular/cdk`.
-    await linker.main(['manifest.json'], new linker.Runfiles({
+    await linker.main(['manifest.json'], new Runfiles({
       // This test assumes an environment where runfiles are not symlinked. Hence
       // we pass a runfile manifest file.
       'RUNFILES_MANIFEST_FILE': 'runfiles.mf',
@@ -487,7 +494,7 @@ describe('link_node_modules', () => {
     });
 
     // TODO(alexeagle): test should control the environment, not just pass through
-    await linker.main(['manifest.json'], new linker.Runfiles(process.env));
+    await linker.main(['manifest.json'], new Runfiles(process.env));
 
     // The linker expects to run as its own process, so it changes the wd
     process.chdir(path.join());
@@ -520,7 +527,7 @@ describe('link_node_modules', () => {
     });
 
     // TODO(alexeagle): test should control the environment, not just pass through
-    await linker.main(['manifest.json'], new linker.Runfiles(process.env));
+    await linker.main(['manifest.json'], new Runfiles(process.env));
 
     // The linker expects to run as its own process, so it changes the wd
     process.chdir(path.join());
@@ -553,7 +560,7 @@ describe('link_node_modules', () => {
     });
 
     // TODO(alexeagle): test should control the environment, not just pass through
-    await linker.main(['manifest.json'], new linker.Runfiles(process.env));
+    await linker.main(['manifest.json'], new Runfiles(process.env));
 
     // The linker expects to run as its own process, so it changes the wd
     process.chdir(path.join());
@@ -579,7 +586,7 @@ describe('link_node_modules', () => {
     });
     writeRunfiles(runfilesManifest);
 
-    await linker.main(['manifest.json'], new linker.Runfiles({
+    await linker.main(['manifest.json'], new Runfiles({
       'RUNFILES_MANIFEST_FILE': 'runfiles.mf',
     }));
