+++ packages/runfiles/package.json
@@ -1,0 +1,22 @@
+{
+    "name": "@bazel/runfiles",
+    "description": "NodeJS Runfile helpers for Bazel",
+    "license": "Apache-2.0",
+    "version": "0.0.0-PLACEHOLDER",
+    "repository": {
+        "type" : "git",
+        "url" : "https://github.com/bazelbuild/rules_nodejs.git",
+        "directory": "packages/runfiles"
+    },
+    "bugs": {
+        "url": "https://github.com/bazelbuild/rules_nodejs/issues"
+    },
+    "keywords": [
+        "bazel",
+        "runfiles",
+        "runfiles helpers"
+    ],
+    "main": "index.js",
+    "types": "index.d.ts",
+    "dependencies": {}
+}
