+++ packages/runfiles/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "module": "commonjs",
+    "target": "ES2015",
+    "types": ["node"]
+  },
+}
